CreatesValidGameBoard fixture in GameBoardBuilderTest compiles
