+++ src/core/builders/builders.cpp
@@ -2,11 +2,12 @@
 
 namespace gameboard {
 
-std::shared_ptr<GameBoardForConcepts> GameBoardBuilder::build(
-    const BoardMapInt_t starting_board
-) {
-  return std::make_shared<GameBoardForConcepts>(starting_board);
-}
+
+// std::shared_ptr<GameBoardForConcepts> GameBoardBuilder::build(
+//     const BoardMapInt_t starting_board
+// ) {
+//   return std::make_shared<GameBoardForConcepts>(starting_board);
+// }
 
 } // namespace gameboard
